Created a post creation component
